+++ cosmogrb/test/test_gbm_background.py
@@ -1,4 +1,4 @@
-from cosmogrb.sampler.background import GBMBackground
+from cosmogrb.instruments.gbm import GBMBackground
 
 
 allowed_dets = (
